Refactor `list` methods of app into module - Integrated `#welcome_commands` into main method - Refactored all `#list` methods of `App` class into modules
